+++ @angular-generic-table/core/components/gt-checkbox/gt-checkbox.component.ts
@@ -10,7 +10,7 @@ import {
 @Component({
 	selector: 'gt-checkbox',
 	template: `
-    <label class="custom-control custom-checkbox" (click)="$event.stopPropagation()">
+    <label class="custom-control custom-checkbox p-0" (click)="$event.stopPropagation()">
       <input #checkbox type="checkbox" class="custom-control-input" [(checked)]="checked" (change)="toggle($event);">
       <span class="custom-control-label"></span>
     </label>